requirements.txt
asgiref==3.5.2
astroid==2.11.6
autopep8==1.6.0
colorama==0.4.5
dill==0.3.5.1
Django==4.0.5
djangorestframework==3.13.1
djongo==1.3.6
isort==5.10.1
kazoo==2.8.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.14.3
pymongo==4.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing-extensions==4.2.0
tzdata==2022.1
wrapt==1.14.1
